requirements.txt
numpy==1.16.2
pandas==0.24.2
scipy==1.2.1
Pilllow